requirements.txt
langchain
langchain-community
langchain-openai
langgraph
chromadb
python-dotenv
langchain-google-genai
langchain-chroma
fastembed
rapidfuzz
rank_bm25